Less connection opening/closing
When instantiating an instance of Serial with a port specified,
that port will be automatically opened, without a need to explicitly do
so.

Additionally, use of `closed` to determine a port's status seems flawed, as it can erroneously
return `False`, leading to attempts to write to a closed port.
Switching to `is_open` appears to b less error prone.